Organism data

Load or build the organism specific KEGG datafor enrichment analysis. This provides a convenience wrapper around FELLA::loadKEGGdata, FELLA::buildGraphFromKEGGREST and FELLA::buildDataFromGraph. See the documentation of these functions for further information.

Usage

organismData(
  organism,
  methods = availableMethods(),
  filter_path = NULL,
  database_directory = organism,
  internal_directory = TRUE,
  damping_factor = 0.85,
  niter = 100
)

Arguments

organism

the KEGG code for the organism of interest

methods

the enrichment techniques to build. Any returned by availableMethods.

filter_path

argument to pass to argument filter.path of FELLA::buildGraphFromKEGGREST. A vector of regular expressions to match pathways to exclude.

database_directory

argument to pass to argument databaseDir of FELLA::buildDataFromGraph and FELLA::loadKEGGdata. The directory name/path in which to save the KEGG data.

internal_directory

logical. Argument to pass to argument internalDir of FELLA::buildDataFromGraph and FELLA::loadKEGGdata. Should the save directory be internal to the FELLA package directory?

damping_factor

argument to pass to argument dampingFactor of FELLA::buildDataFromGraph. A value between 0 and 1 for the PageRank damping factor.

niter

argument to pass to argument niter of FELLA::buildDataFromGraph. A value between 10 and 1000. The number of iterations to estimate the values for CC size.

Value

An object of S4 class FELLA.DATA containing the KEGG data for the specified organism.

Examples

## Load the example organism data available from within the package
organismData(
  'bdi',
  database_directory = system.file(
    'bdi',
    package = 'riches'),
  internal_directory = FALSE
)
#> Loading KEGG graph data...
#> Done.
#> Loading hypergeom data...
#> Loading matrix...
#> Done.
#> Loading diffusion data...
#> Loading matrix...
#> 'diffusion.matrix.RData' not present in:/home/runner/work/_temp/Library/riches/bdi/diffusion.matrix.RData. Simulated permutations may execute slower for diffusion.
#> Done.
#> Loading rowSums...
#> 'diffusion.rowSums.RData' not present in:/home/runner/work/_temp/Library/riches/bdi/diffusion.rowSums.RData. Z-scores won't be available for diffusion.
#> Done.
#> Loading pagerank data...
#> Loading matrix...
#> 'pagerank.matrix.RData' not present in:/home/runner/work/_temp/Library/riches/bdi/pagerank.matrix.RData. Simulated permutations may execute slower for pagerank.
#> Done.
#> Loading rowSums...
#> 'pagerank.rowSums.RData' not present in:/home/runner/work/_temp/Library/riches/bdi/pagerank.rowSums.RData. Z-scores won't be available for pagerank.
#> Done.
#> Data successfully loaded.
#> General data:
#> - KEGG graph:
#>   * Nodes:  13085 
#>   * Edges:  42910 
#>   * Density:  0.0002506364624 
#>   * Categories:
#>     + pathway [136]
#>     + module [218]
#>     + enzyme [993]
#>     + reaction [6703]
#>     + compound [5035]
#>   * Size:  7.1 Mb 
#> - KEGG names are ready.
#> -----------------------------
#> Hypergeometric test:
#> - Matrix is ready
#>   * Dim:  5035 x 136 
#>   * Size:  473.8 Kb
#> -----------------------------
#> Heat diffusion:
#> - Matrix not loaded.
#> - RowSums not loaded.
#> -----------------------------
#> PageRank:
#> - Matrix not loaded.
#> - RowSums not loaded.